Remove puzzle input from 2023-Day-2-Carbon
As mentioned in the previous commit, I didn't realize that we're asked
not to put the puzzle inputs in our repositories, so I'm removing them
from mine. Apologies for the oversight.

Since I had to solve this puzzle rather hackily with Carbon, removing
the puzzle input was more involved for this day